Remove inline test modules from BGP source files
- Remove test modules from attributes.rs, capabilities.rs, codec.rs, messages.rs, nlri.rs
- Eliminates 47 redundant inline tests while preserving comprehensive separate test files
- Source files now focus purely on implementation without embedded test code
