+++ src/bgp/messages.rs
@@ -10,7 +10,6 @@ use std::io::prelude::*;
 use std::io::Cursor;
 use std::mem::size_of;
 use std::net::IpAddr;
-use std::net::Ipv4Addr;
 
 use super::attributes::*;
 use super::capabilities::*;
@@ -222,8 +221,7 @@ impl TryFrom<Vec<u8>> for BGPUpdateMessage {
                 break;
             }
             let plen = src[i];
-            validate_prefix_length_afi(plen, &Afi::Ipv4).map_err(|e| BgpError::Message(e.to_string()))?;
-            let plen_bytes = prefix_bytes(plen);
+            let plen_bytes = prefix_bytes(plen, &Afi::Ipv4).map_err(|e| BgpError::Message(e.to_string()))?;
             let end = i.checked_add(plen_bytes + 1)
                 .ok_or_else(|| BgpError::Message("Buffer offset overflow".to_string()))?;
             if end > src.len() {
@@ -304,8 +302,7 @@ impl TryFrom<Vec<u8>> for BGPUpdateMessage {
         let mut routes: Vec<Nlri> = vec![];
         while i < total_len {
             let plen = src[i];
-            validate_prefix_length_afi(plen, &Afi::Ipv4).map_err(|e| BgpError::Message(e.to_string()))?;
-            let plen_bytes = prefix_bytes(plen);
+            let plen_bytes = prefix_bytes(plen, &Afi::Ipv4).map_err(|e| BgpError::Message(e.to_string()))?;
             let end = i.checked_add(plen_bytes + 1)
                 .ok_or_else(|| BgpError::Message("Buffer offset overflow".to_string()))?;
             let buf = Ipv4Octets {
@@ -514,196 +511,3 @@ impl Message {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::neighbor::Capabilities;
-
-    #[test]
-    fn test_bgp_open_message_new() {
-        let caps = Capabilities {
-            multiprotocol: Some(vec![AddressFamily {
-                afi: Afi::Ipv4,
-                safi: Safi::NLRIUnicast,
-            }]),
-            ..Default::default()
-        };
-        let open = BGPOpenMessage::new(65000, 16843009, 180, caps).unwrap();
-        assert_eq!(open.version, VERSION);
-        assert_eq!(open.asn, 65000);
-        assert_eq!(open.hold_time, 180);
-        assert_eq!(open.router_id, 16843009);
-    }
-
-    #[test]
-    fn test_bgp_open_message_serialization() {
-        let caps = Capabilities::default();
-        let open = BGPOpenMessage::new(65000, 16843009, 180, caps).unwrap();
-        let bytes: Vec<u8> = open.clone().into();
-        let parsed: BGPOpenMessage = bytes.into();
-
-        assert_eq!(parsed.version, open.version);
-        assert_eq!(parsed.asn, open.asn);
-        assert_eq!(parsed.hold_time, open.hold_time);
-        assert_eq!(parsed.router_id, open.router_id);
-    }
-
-    #[test]
-    fn test_bgp_update_message_empty() {
-        let update = BGPUpdateMessage::new().unwrap();
-        assert!(update.withdrawn_routes.is_empty());
-        assert!(update.path_attributes.is_empty());
-        assert!(update.nlri.is_empty());
-    }
-
-    #[test]
-    fn test_bgp_update_message_with_routes() {
-        let nlri1 = Nlri {
-            net: "10.0.0.0/24".parse().unwrap(),
-        };
-        let nlri2 = Nlri {
-            net: "10.1.0.0/24".parse().unwrap(),
-        };
-
-        let attr = PathAttribute::origin(OriginType::Igp);
-
-        let update = BGPUpdateMessageBuilder::default()
-            .withdrawn_routes(vec![nlri1])
-            .path_attributes(vec![attr])
-            .nlri(vec![nlri2])
-            .build()
-            .unwrap();
-
-        assert_eq!(update.withdrawn_routes.len(), 1);
-        assert_eq!(update.path_attributes.len(), 1);
-        assert_eq!(update.nlri.len(), 1);
-    }
-
-    #[test]
-    fn test_bgp_update_message_serialization() {
-        let nlri = Nlri {
-            net: "192.0.2.0/24".parse().unwrap(),
-        };
-        let attrs = vec![
-            PathAttribute::origin(OriginType::Igp),
-            PathAttribute::aspath(vec![ASPATHSegment {
-                segment_type: ASPATHSegmentType::AsSequence,
-                as_list: vec![65000],
-            }]),
-            PathAttribute::nexthop(Ipv4Addr::new(192, 0, 2, 1)),
-        ];
-
-        let update = BGPUpdateMessageBuilder::default()
-            .withdrawn_routes(vec![])
-            .path_attributes(attrs)
-            .nlri(vec![nlri])
-            .build()
-            .unwrap();
-
-        let bytes: Vec<u8> = update.clone().into();
-        let parsed: BGPUpdateMessage = bytes
-            .try_into()
-            .expect("Failed to parse BGP update message");
-
-        assert_eq!(parsed.withdrawn_routes.len(), 0);
-        assert_eq!(parsed.path_attributes.len(), 3);
-        assert_eq!(parsed.nlri.len(), 1);
-    }
-
-    #[test]
-    fn test_bgp_notification_message() {
-        let notif = BGPNotificationMessage::new(ErrorCode::UpdateMessage, 3).unwrap();
-        assert_eq!(notif.error_code, ErrorCode::UpdateMessage);
-        assert_eq!(notif.error_subcode, 3);
-        assert!(notif.data.is_empty());
-    }
-
-    #[test]
-    fn test_bgp_notification_message_serialization() {
-        let notif = BGPNotificationMessageBuilder::default()
-            .error_code(ErrorCode::HoldTimerExpired)
-            .error_subcode(0)
-            .data(vec![1, 2, 3])
-            .build()
-            .unwrap();
-
-        let bytes: Vec<u8> = notif.clone().into();
-        assert_eq!(bytes[0], ErrorCode::HoldTimerExpired as u8);
-        assert_eq!(bytes[1], 0);
-        assert_eq!(&bytes[2..], &[1, 2, 3]);
-
-        let parsed: BGPNotificationMessage = bytes[0..2].to_vec().into();
-        assert_eq!(parsed.error_code, ErrorCode::HoldTimerExpired);
-        assert_eq!(parsed.error_subcode, 0);
-    }
-
-    #[test]
-    fn test_bgp_keepalive_message() {
-        let keepalive = BGPKeepaliveMessage::new().unwrap();
-        assert_eq!(keepalive.byte_len(), 0);
-
-        let bytes: Vec<u8> = keepalive.into();
-        assert!(bytes.is_empty());
-    }
-
-    #[test]
-    fn test_message_complete_serialization() {
-        let body = BGPKeepaliveMessage::new().unwrap();
-        let msg = Message::new(MessageType::Keepalive, BGPMessageBody::Keepalive(body)).unwrap();
-
-        let serialized_bytes: Vec<u8> = msg.into();
-        assert_eq!(serialized_bytes[0], MessageType::Keepalive as u8);
-    }
-
-    #[test]
-    fn test_message_from_bytes() {
-        // Create a complete BGP message with marker, length, and type
-        let mut msg_bytes = vec![];
-        msg_bytes.extend_from_slice(&MARKER); // Marker
-        msg_bytes.extend_from_slice(&[0, 19]); // Length
-        msg_bytes.push(MessageType::Keepalive as u8); // Type
-
-        let msg: Message = msg_bytes.try_into().expect("Failed to parse BGP message");
-        assert_eq!(msg.header.message_type, MessageType::Keepalive);
-    }
-
-    #[test]
-    fn test_prefix_validation() {
-        use crate::bgp::types::*;
-        
-        // Valid IPv4 prefix lengths
-        assert!(validate_prefix_length_afi(0, &Afi::Ipv4).is_ok());
-        assert!(validate_prefix_length_afi(24, &Afi::Ipv4).is_ok());
-        assert!(validate_prefix_length_afi(32, &Afi::Ipv4).is_ok());
-
-        // Invalid IPv4 prefix lengths
-        assert!(validate_prefix_length_afi(33, &Afi::Ipv4).is_err());
-        assert!(validate_prefix_length_afi(255, &Afi::Ipv4).is_err());
-        
-        // Valid IPv6 prefix lengths
-        assert!(validate_prefix_length_afi(64, &Afi::Ipv6).is_ok());
-        assert!(validate_prefix_length_afi(128, &Afi::Ipv6).is_ok());
-        
-        // Invalid IPv6 prefix lengths
-        assert!(validate_prefix_length_afi(129, &Afi::Ipv6).is_err());
-    }
-
-    #[test]
-    fn test_bgp_update_malicious_prefix_length() {
-        // Create malicious BGP UPDATE with invalid prefix length
-        let mut malicious_update = vec![];
-        malicious_update.extend_from_slice(&[0, 0]); // No withdrawn routes
-        malicious_update.extend_from_slice(&[0, 0]); // No path attributes
-        malicious_update.push(255); // Invalid prefix length (> 32)
-        malicious_update.extend_from_slice(&[1, 2, 3, 4]); // Some octets
-
-        let result: Result<BGPUpdateMessage, BgpError> = malicious_update.try_into();
-        assert!(result.is_err());
-        
-        if let Err(BgpError::Message(msg)) = result {
-            assert!(msg.contains("Invalid NLRI prefix length"));
-        } else {
-            panic!("Expected BgpError::Message");
-        }
-    }
-}
