+++ src/bgp/messages_tests.rs
@@ -614,7 +614,7 @@ fn test_message_all_types_valid() {
 #[test]
 fn test_message_round_trip_serialization_valid() {
     let original_body = BGPKeepaliveMessage::new().unwrap();
-    let original_msg = Message::new(MessageType::Keepalive, BGPMessageBody::Keepalive(original_body)).unwrap();
+    let _original_msg = Message::new(MessageType::Keepalive, BGPMessageBody::Keepalive(original_body)).unwrap();
     
     // Create a complete BGP message manually (like the working test in messages.rs)
     let mut complete_msg = vec![];
