+++ src/neighbor/fsm.rs
@@ -15,7 +15,7 @@ use tokio::sync::mpsc;
 use tokio::sync::Mutex;
 use tokio_util::codec::Framed;
 
-fn calculate_retry_delay(
+pub(super) fn calculate_retry_delay(
     base_retry_time: u16,
     retry_counter: usize,
     exponential_backoff: bool,
@@ -30,7 +30,14 @@ fn calculate_retry_delay(
     // Add 10% jitter to prevent thundering herd
     let jitter_range = base_delay / 10;
     let jitter = (retry_counter as u64 * 1237) % (jitter_range + 1); // Pseudo-random
-    base_delay + jitter
+    let total_delay = base_delay + jitter;
+    
+    // Ensure we never exceed 3600 seconds (1 hour) for exponential backoff
+    if exponential_backoff {
+        total_delay.min(3600)
+    } else {
+        total_delay
+    }
 }
 
 pub async fn init_peer(n: Arc<Mutex<BGPNeighbor>>) {
@@ -54,7 +61,7 @@ pub async fn connect(
         log::debug!("Neighbor ribtx set: {:?}", n.ribtx.keys());
     }
     // Extract remote address, local_ip, and local_asn from the neighbor
-    let (remote_addr, local_ip, local_asn) = {
+    let (remote_addr, _local_ip, _local_asn) = {
         let n = neighbor.lock().await;
         let remote_ip = n
             .remote_ip
@@ -512,33 +519,3 @@ pub async fn process_event_established(
     Ok(())
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_calculate_retry_delay_fixed() {
-        let delay1 = calculate_retry_delay(10, 0, false);
-        let delay2 = calculate_retry_delay(10, 3, false);
-        
-        assert!(delay1 >= 10 && delay1 <= 11);
-        assert!(delay2 >= 10 && delay2 <= 11);
-    }
-
-    #[test]
-    fn test_calculate_retry_delay_exponential() {
-        let delay1 = calculate_retry_delay(5, 0, true);
-        let delay2 = calculate_retry_delay(5, 1, true);
-        let delay3 = calculate_retry_delay(5, 2, true);
-        
-        assert!(delay1 >= 5 && delay1 <= 5);
-        assert!(delay2 >= 10 && delay2 <= 11);
-        assert!(delay3 >= 20 && delay3 <= 22);
-    }
-
-    #[test]
-    fn test_exponential_backoff_caps_at_max() {
-        let delay = calculate_retry_delay(60, 20, true);
-        assert!(delay <= 3600);
-    }
-}
